Restore no camera errors
