+++ src/page/template/list/conversations.htm
@@ -27,7 +27,8 @@
           conversation: getConversationById(activeCall.conversationId),
           multitasking: multitasking,
           callingRepository: callingRepository,
-          callActions: callActions"
+          callActions: callActions,
+          hasAccessToCamera: hasAccessToCamera()"
         data-uie-name="item-call">
       </conversation-list-calling-cell>
     <!-- /ko -->
