+++ src/script/components/calling/fullscreenVideoCall.js
@@ -119,120 +119,6 @@ export class FullscreenVideoCalling {
       startedAtSubscription.dispose();
       gridSubscription.dispose();
     };
-
-    /*
-    this.clickedOnCancelScreen = this.clickedOnCancelScreen.bind(this);
-    this.clickedOnChooseScreen = this.clickedOnChooseScreen.bind(this);
-    this.chooseSharedScreen = this.chooseSharedScreen.bind(this);
-
-    this.callingRepository = repositories.calling;
-    this.conversationRepository = repositories.conversation;
-    this.mediaRepository = repositories.media;
-    this.userRepository = repositories.user;
-
-    this.contentViewModel = mainViewModel.content;
-    this.multitasking = this.contentViewModel.multitasking;
-
-    this.hasSelfVideo = this.streamHandler.hasActiveVideo;
-    this.selfStreamState = this.streamHandler.selfStreamState;
-    this.localVideoStream = this.streamHandler.localMediaStream;
-    this.remoteVideoStreamsInfo = this.streamHandler.remoteMediaStreamInfoIndex.video;
-
-    this.minimizeTimeout = undefined;
-
-    this.calls = this.callingRepository.calls;
-    this.joinedCall = this.callingRepository.joinedCall;
-
-    this.remoteUser = ko.pureComputed(() => {
-      const [participantEntity] = this.joinedCall() ? this.joinedCall().participants() : [];
-
-      if (participantEntity) {
-        return participantEntity.user;
-      }
-    });
-
-    this.showRemoteParticipant = ko.pureComputed(() => {
-      const showRemoteParticipant = this.remoteUser() && !this.multitasking.isMinimized() && !this.isChoosingScreen();
-      return showRemoteParticipant && this.isCallOngoing() && !this.showRemoteVideo();
-    });
-
-    this.showRemoteVideo = ko.pureComputed(() => {
-      if (this.isCallOngoing()) {
-        const remoteVideoState = this.joinedCall() && this.joinedCall().isRemoteVideoCall();
-        return remoteVideoState && this.remoteVideoStreamsInfo().length;
-      }
-    });
-
-    this.showSwitchCamera = ko.pureComputed(() => {
-      const hasMultipleCameras = this.availableDevices.videoInput().length > 1;
-      const isVisible = hasMultipleCameras && this.localVideoStream() && this.selfStreamState.videoSend();
-      return this.isCallOngoing() && isVisible;
-    });
-    this.showSwitchScreen = ko.pureComputed(() => {
-      const hasMultipleScreens = this.availableDevices.screenInput().length > 1;
-      const isVisible = hasMultipleScreens && this.localVideoStream() && this.selfStreamState.screenSend();
-      return this.isCallOngoing() && isVisible;
-    });
-
-    this.showToggleVideo = ko.pureComputed(() => {
-      return this.joinedCall() ? this.joinedCall().conversationEntity.supportsVideoCall(false) : false;
-    });
-    this.disableToggleScreen = ko.pureComputed(() => !this.callingRepository.supportsScreenSharing);
-
-    this.visibleCallId = undefined;
-    this.joinedCall.subscribe(callEntity => {
-      if (callEntity) {
-        const isVisibleId = this.visibleCallId === callEntity.id;
-        if (!isVisibleId) {
-          this.visibleCallId = callEntity.id;
-
-          // FIXME find a better condition to actually minimize/maximize the call
-          // we should do this when we check that everything is alright with audio calls also
-          if (this.showRemoteVideo()) {
-            this.multitasking.isMinimized(false);
-            return this.logger.info(`Maximizing video call '${callEntity.id}' to full-screen`, callEntity);
-          }
-
-          //this.multitasking.isMinimized(true);
-          this.logger.info(`Minimizing audio call '${callEntity.id}' from full-screen`, callEntity);
-        }
-      } else {
-        this.visibleCallId = undefined;
-        this.multitasking.autoMinimize(true);
-        this.multitasking.isMinimized(false);
-        this.logger.info('Resetting full-screen calling to maximize');
-      }
-    });
-
-    this.showRemoteParticipant.subscribe(showRemoteParticipant => {
-      if (this.minimizeTimeout) {
-        window.clearTimeout(this.minimizeTimeout);
-        this.minimizeTimeout = undefined;
-      }
-
-      const isVideoCall = showRemoteParticipant && this.videodCall();
-      const shouldAutoMinimize = isVideoCall && this.multitasking.autoMinimize() && !this.isChoosingScreen();
-      if (shouldAutoMinimize) {
-        const remoteUserName = this.remoteUser() ? this.remoteUser().name() : '';
-
-        const callId = this.videodCall().id;
-        const logMessage = `Scheduled minimizing call '${callId}' as remote user '${remoteUserName}' is not videod`;
-        this.logger.info(logMessage);
-        this.minimizeTimeout = window.setTimeout(() => {
-          if (!this.isChoosingScreen()) {
-            this.multitasking.isMinimized(true);
-          }
-          const message = `Minimizing call '${callId}' on timeout as remote user '${remoteUserName}' is not videod`;
-          this.logger.info(message);
-        }, VideoCallingViewModel.CONFIG.AUTO_MINIMIZE_TIMEOUT);
-      }
-    });
-
-
-    amplify.subscribe(WebAppEvents.CALL.MEDIA.CHOOSE_SCREEN, this.chooseSharedScreen);
-
-    this.MediaDeviceType = MediaDeviceType;
-    */
   }
 
   chooseSharedScreen(conversationId) {
